Move text components to separate folders
