+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/body-small-semi-bold/index.ts
@@ -1,0 +1,1 @@
+export * from './BodySmallSemiBold';
