+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/index.ts
@@ -1,0 +1,8 @@
+export * from './body-small';
+export * from './body-small-semi-bold';
+export * from './footnote-mini';
+export * from './label-tiny';
+export * from './label-tiny-semi-bold';
+export * from './title-big';
+export * from './title-large';
+export * from './title-medium';
