+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/label-tiny-semi-bold/LabelTinySemiBold.tsx
@@ -1,4 +1,4 @@
-import { Text, TextProps } from './Text';
+import { Text, TextProps } from '../../Text';
 
 export type LabelTinySemiBoldProps<E extends React.ElementType = 'span'> = TextProps<E>;
