+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/label-tiny-semi-bold/index.ts
@@ -1,0 +1,1 @@
+export * from './LabelTinySemiBold';
