+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/label-tiny/index.ts
@@ -1,0 +1,1 @@
+export * from './LabelTiny';
