+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/title-large/index.ts
@@ -1,0 +1,1 @@
+export * from './TitleLarge';
