+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/title-medium/TitleMedium.tsx
@@ -1,4 +1,4 @@
-import { Text, TextProps } from './Text';
+import { Text, TextProps } from '../../Text';
 
 export type TitleMediumProps<E extends React.ElementType = 'span'> = TextProps<E>;
