+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/components/title-medium/index.ts
@@ -1,0 +1,1 @@
+export * from './TitleMedium';
