+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/index.ts
@@ -1,9 +1,2 @@
-export * from './BodySmall';
-export * from './BodySmallSemiBold';
-export * from './FootnoteMini';
-export * from './LabelTiny';
-export * from './LabelTinySemiBold';
 export * from './Text';
-export * from './TitleBig';
-export * from './TitleLarge';
-export * from './TitleMedium';
+export * from './components';
